allow decimal precision on net income

+++ src/API/app/domain/models/income_period.py
@@ -8,5 +8,5 @@ class IncomePeriod:
     id: str
     start_date: str
     yearly_gross_amount: int
-    net_percentage: int
+    net_percentage: float
     end_date: str | None = None
